Moved readline frontend to its own script.
